Wrap long lines in descriptions (#325)
Wrap lines, it's more suited to a plaintext format like markdown.

For rendering, paragraphs don't _need_ to be wrapped. But markdown uses
a double-linebreak as paragraph separator so that paragraphs _can_ be
wrapped. As Gruber writes:

    Readability, however, is emphasized above all else. A
    Markdown-formatted document should be publishable as-is, as plain
    text, without looking like it’s been marked up with tags or
    formatting instructions.

For me, this includes wrapping lines. A lot of source code, such as
C or javascript, doesn't need wrapping either. But yet we do, for
readability and better diffs.

Here's a markdown style guide that raises the same points (both pro
and con):
http://www.cirosantilli.com/markdown-style-guide/#line-wrapping